+++ docs/changes.rst
@@ -30,6 +30,7 @@ Not yet released.
 * Display addon parameters on overview.
 * Sentry is now supported through modern Sentry SDK instead of Raven.
 * Changed example settings to be better fit for production environment.
+* Added automated backups using Borg backup.
 
 weblate 3.8
 -----------